The language browser and the language selection page are finally working (I believe)
1. Now the language browser does not create a new identical page unnecessarily.

2. Both pages now have a scroll bar. It is convenient, especially when the mouse wheel is broken.....

3.1 The button for opening downloaded languages in the explorer and the button for opening the language page of the browser have also been updated. Now all the buttons are created by “copying” the "General" button from the navigation bar.

3.2 However, this leads to a small uncritical graphical error. All the buttons in the language browser look like a “Pressed button” (a button highlighted in white). To be honest, I do not know how to fix it. And whether it's worth fixing at all.

Possible problems:

For the test, I downloaded almost all available languages into the language browser. When I reloaded the menu (entered and exited the level), some languages were missing. Even when I moved the en-GB file from the build to the translation folder, it was still not visible. I do not know if it is my fault or the mod itself, but I will have to test it again sometime.

Any ideas for improvement:

Most likely, we should move the process of creating buttons for installed languages into a separate method (or function, I get confused all the time), where first all the buttons except “openLangFolder” and “LangBrowser” are deleted, and then the buttons are created anew. This would allow calling this function and forcibly updating the list of local languages (when clicking on the ”Refresh" button on the "Downloaded languages" page or after downloading a language).